faiss_api/requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
